# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gbolix API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List current user's projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (service request)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (client)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # ── Files ─────────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: userType
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get detailed user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{id}/role:
    patch:
      operationId: adminChangeUserRole
      tags: [admin]
      summary: Change a user's role (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleChangeInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/deactivate:
    post:
      operationId: adminDeactivateUser
      tags: [admin]
      summary: Deactivate or reactivate a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
              required:
                - isActive
      responses:
        "200":
          description: Updated
...
  /admin/projects:
    get:
      operationId: adminListProjects
      tags: [admin]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminProject"
  /admin/projects/{id}:
    patch:
      operationId: adminUpdateProject
      tags: [admin]
      summary: Update a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProject"
  /admin/projects/{id}/start-conversation:
    post:
      operationId: adminStartConversation
      tags: [admin]
      summary: Start a conversation on a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProject"
  /admin/projects/{id}/assign:
    post:
      operationId: adminAssignFreelancer
      tags: [admin]
      summary: Assign or unassign freelancers to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignFreelancerInput"
      responses:
        "200":
          description: Updated assignments
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectId:
                    type: integer
                  freelancerIds:
                    type: array
                    items:
                      type: integer
                required:
                  - projectId
                  - freelancerIds
...
components:
  schemas:
...
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [request_submitted, admin_response, file_uploaded, status_change]
        description:
          type: string
        createdAt:
          type: string
        projectId:
          type: ["integer", "null"]
        projectTitle:
          type: ["string", "null"]
      required:
        - id
        - type
        - description
        - createdAt
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceType:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [submitted, queued, in_progress, review, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        price:
          type: ["number", "null"]
        hasConversation:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - serviceType
        - title
        - description
        - status
        - priority
        - hasConversation
        - createdAt
        - updatedAt
    ProjectInput:
      type: object
      properties:
        serviceType:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        price:
          type: number
      required:
        - serviceType
        - title
        - description
        - priority
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        userType:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        companySize:
          type: ["string", "null"]
        acquisitionSource:
          type: ["string", "null"]
        registrationDate:
          type: string
        totalRequests:
          type: integer
        role:
          type: string
          enum: [owner, admin, freelancer, client]
        isActive:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        lastLoginAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - email
        - registrationDate
        - totalRequests
        - role
        - isActive
    AdminUserDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        userType:
          type: ["string", "null"]
        companySize:
          type: ["string", "null"]
        acquisitionSource:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin, freelancer, client]
        isActive:
          type: boolean
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        lastLoginAt:
          type: ["string", "null"]
        totalProjects:
          type: integer
        completedProjects:
          type: integer
        totalFiles:
          type: integer
        totalMessages:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
        - totalProjects
        - completedProjects
        - totalFiles
        - totalMessages
        - recentActivity
    RoleChangeInput:
      type: object
      properties:
        role:
          type: string
          enum: [owner, admin, freelancer, client]
      required:
        - role
...
    AssignFreelancerInput:
      type: object
      properties:
        freelancerIds:
          type: array
          items:
            type: integer
      required:
        - freelancerIds
    AdminProject:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        serviceType:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [submitted, queued, in_progress, review, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        price:
          type: ["number", "null"]
        internalNotes:
          type: ["string", "null"]
        hasConversation:
          type: boolean
        clientName:
          type: string
        clientEmail:
          type: string
        assignedFreelancers:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
            required:
              - id
              - name
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - serviceType
        - description
        - status
        - priority
        - hasConversation
        - clientName
        - clientEmail
        - assignedFreelancers
        - createdAt
        - updatedAt
    AdminProjectUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, queued, in_progress, review, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        internalNotes:
          type: string
        price:
          type: number
...